requirements.txt
streamlit
chromadb
python-dotenv
PyPDF2
supabase
sentence-transformers
groq
reportlab
